requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==3.0.4
chardet2==2.0.3
charset-normalizer==2.0.9
contextlib2==21.6.0
idna==2.7
lxml==4.7.1
pathlib2==2.3.6
pyTelegramBotAPI==4.2.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.20.1
schedule==1.1.0
six==1.16.0
soupsieve==2.3.1
urllib3==1.24.3
yattag==1.14.0
